requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.3.1
APScheduler==3.9.1.post1
apscheduler-di==0.0.7
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
betterlogging==0.2.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.7
cycler==0.11.0
environs==9.5.0
fonttools==4.39.2
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
httplib2==0.21.0
idna==3.4
kiwisolver==1.4.4
magic-filter==1.0.9
marshmallow==3.19.0
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
packaging==22.0
Pillow==9.4.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rodi==1.1.3
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.45
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
yarl==1.8.2
